+++ src/model/data.ts
@@ -15,10 +15,8 @@ export const products: Products = [
     thumbnail: 'https://i.dummyjson.com/data/products/1/thumbnail.jpg',
     images: [
       'https://i.dummyjson.com/data/products/1/1.jpg',
-      'https://i.dummyjson.com/data/products/1/2.jpg',
       'https://i.dummyjson.com/data/products/1/3.jpg',
       'https://i.dummyjson.com/data/products/1/4.jpg',
-      'https://i.dummyjson.com/data/products/1/thumbnail.jpg',
     ],
   },
   {
@@ -177,7 +175,6 @@ export const products: Products = [
     thumbnail: 'https://i.dummyjson.com/data/products/10/thumbnail.jpeg',
     images: [
       'https://i.dummyjson.com/data/products/10/1.jpg',
-      'https://i.dummyjson.com/data/products/10/2.jpg',
       'https://i.dummyjson.com/data/products/10/3.jpg',
       'https://i.dummyjson.com/data/products/10/thumbnail.jpeg',
     ],
@@ -291,7 +288,6 @@ export const products: Products = [
       'https://i.dummyjson.com/data/products/16/2.webp',
       'https://i.dummyjson.com/data/products/16/3.jpg',
       'https://i.dummyjson.com/data/products/16/4.jpg',
-      'https://i.dummyjson.com/data/products/16/thumbnail.jpg',
     ],
   },
   {
@@ -364,7 +360,6 @@ export const products: Products = [
       'https://i.dummyjson.com/data/products/20/1.jpg',
       'https://i.dummyjson.com/data/products/20/2.jpg',
       'https://i.dummyjson.com/data/products/20/3.jpg',
-      'https://i.dummyjson.com/data/products/20/4.jpg',
       'https://i.dummyjson.com/data/products/20/thumbnail.jpg',
     ],
   },
@@ -381,7 +376,6 @@ export const products: Products = [
     thumbnail: 'https://i.dummyjson.com/data/products/21/thumbnail.png',
     images: [
       'https://i.dummyjson.com/data/products/21/1.png',
-      'https://i.dummyjson.com/data/products/21/2.jpg',
       'https://i.dummyjson.com/data/products/21/3.jpg',
     ],
   },
@@ -418,7 +412,6 @@ export const products: Products = [
       'https://i.dummyjson.com/data/products/23/2.jpg',
       'https://i.dummyjson.com/data/products/23/3.jpg',
       'https://i.dummyjson.com/data/products/23/4.jpg',
-      'https://i.dummyjson.com/data/products/23/thumbnail.jpg',
     ],
   },
   {
@@ -436,7 +429,6 @@ export const products: Products = [
       'https://i.dummyjson.com/data/products/24/1.jpg',
       'https://i.dummyjson.com/data/products/24/2.jpg',
       'https://i.dummyjson.com/data/products/24/3.jpg',
-      'https://i.dummyjson.com/data/products/24/4.jpg',
       'https://i.dummyjson.com/data/products/24/thumbnail.jpg',
     ],
   },
@@ -475,8 +467,6 @@ export const products: Products = [
       'https://i.dummyjson.com/data/products/26/2.jpg',
       'https://i.dummyjson.com/data/products/26/3.jpg',
       'https://i.dummyjson.com/data/products/26/4.jpg',
-      'https://i.dummyjson.com/data/products/26/5.jpg',
-      'https://i.dummyjson.com/data/products/26/thumbnail.jpg',
     ],
   },
   {
@@ -532,8 +522,6 @@ export const products: Products = [
       'https://i.dummyjson.com/data/products/29/1.jpg',
       'https://i.dummyjson.com/data/products/29/2.jpg',
       'https://i.dummyjson.com/data/products/29/3.webp',
-      'https://i.dummyjson.com/data/products/29/4.webp',
-      'https://i.dummyjson.com/data/products/29/thumbnail.webp',
     ],
   },
   {
@@ -1217,8 +1205,6 @@ export const products: Products = [
       'https://i.dummyjson.com/data/products/66/1.jpg',
       'https://i.dummyjson.com/data/products/66/2.jpg',
       'https://i.dummyjson.com/data/products/66/3.jpg',
-      'https://i.dummyjson.com/data/products/66/4.JPG',
-      'https://i.dummyjson.com/data/products/66/thumbnail.jpg',
     ],
   },
   {
@@ -1289,7 +1275,6 @@ export const products: Products = [
     images: [
       'https://i.dummyjson.com/data/products/70/1.jpg',
       'https://i.dummyjson.com/data/products/70/2.jpg',
-      'https://i.dummyjson.com/data/products/70/thumbnail.jpg',
     ],
   },
   {
@@ -1415,8 +1400,6 @@ export const products: Products = [
     images: [
       'https://i.dummyjson.com/data/products/77/1.jpg',
       'https://i.dummyjson.com/data/products/77/2.jpg',
-      'https://i.dummyjson.com/data/products/77/3.jpg',
-      'https://i.dummyjson.com/data/products/77/thumbnail.jpg',
     ],
   },
   {
