+++ src/view/headerWiew.ts
@@ -1,4 +1,5 @@
 import { CostAndCount } from '../types/basket';
+import { getCostAndCount } from '../common/basketHelper';
 
 const totalCost: HTMLElement | null = document.getElementById('total-cost');
 const totalCount: HTMLElement | null = document.getElementById('basket-count');
@@ -12,4 +13,8 @@ const headerWiew = ({ cost, count }:CostAndCount):void => {
   }
 };
 
-export default headerWiew;
+const changeHeaderWiew = (): void => {
+  headerWiew(getCostAndCount());
+};
+
+export default changeHeaderWiew;
